+++ src/components/client/client-main.css
@@ -1,29 +1,34 @@
 .Modal {
   position: fixed;
   bottom: -150vh;
-  background-color: #fff;
+  background-color: rgb(175, 204, 244);
   width: 300px;
   left: 0;
   padding: 12px 12px;
   transition: bottom 0.3s ease-out;
   z-index: 10;
+  border-radius: 15px;
 }
 
 .Modal.Show {
   bottom: 0;
 }
 
-.Overlay {
-  background-color: rgba(0, 0, 0, 0.55);
-  top: 0;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  position: fixed;
-  display: none;
-  z-index: 5;
+.close {
+  color: #fff;
+  background-color: #b3b1b1;
+  position: absolute;
+  top: -8px;
+  right: -8px;
+  font-size: 15px;
+  border-radius: 50%;
+  border: 2px solid #333;
 }
 
-.Overlay.Show {
-  display: block;
+.close:hover,
+.close:focus {
+  color: #000;
+  text-decoration: none;
+  cursor: pointer;
+  outline: none;
 }
