--- src/components/client/text-to-speech.css
@@ -1,18 +1,0 @@
-.close {
-  color: #fff;
-  background-color: #999;
-  position: absolute;
-  top: -8px;
-  right: -8px;
-  font-size: 15px;
-  border-radius: 50%;
-  border: 2px solid #333;
-}
-
-.close:hover,
-.close:focus {
-  color: #000;
-  text-decoration: none;
-  cursor: pointer;
-  outline: none;
-}
